+++ acme/acme/client.py
@@ -757,6 +757,13 @@ class ClientV2(ClientBase):
         """
         return self._revoke(cert, rsn, self.directory['revokeCert'])
 
+    def external_account_required(self):
+        """Checks if ACME server requires External Account Binding authentication."""
+        if hasattr(self.directory, 'meta') and self.directory.meta.external_account_required:
+            return True
+        else:
+            return False
+
     def _post_as_get(self, *args, **kwargs):
         """
         Send GET request using the POST-as-GET protocol if needed.
@@ -918,6 +925,15 @@ class BackwardsCompatibleClientV2(object):
             return 2
         else:
             return 1
+
+    def external_account_required(self):
+        """Checks if the server requires an external account for ACMEv2 servers.
+
+        Always return False for ACMEv1 servers, as it doesn't use External Account Binding."""
+        if self.acme_version == 1:
+            return False
+        else:
+            return self.client.external_account_required()
 
 
 class ClientNetwork(object):  # pylint: disable=too-many-instance-attributes